requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
